Move entity database interfaces to db package
