+++ obelisk-server/src/main/java/org/burrow_studios/obelisk/server/db/EntityDatabase.java
@@ -3,14 +3,14 @@ package org.burrow_studios.obelisk.server.db;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import org.burrow_studios.obelisk.server.Main;
-import org.burrow_studios.obelisk.server.its.db.BoardDB;
-import org.burrow_studios.obelisk.server.its.db.IssueState;
-import org.burrow_studios.obelisk.server.moderation.db.project.ProjectDB;
-import org.burrow_studios.obelisk.server.moderation.db.project.ProjectState;
-import org.burrow_studios.obelisk.server.moderation.db.ticket.TicketDB;
-import org.burrow_studios.obelisk.server.moderation.db.ticket.TicketState;
-import org.burrow_studios.obelisk.server.users.db.group.GroupDB;
-import org.burrow_studios.obelisk.server.users.db.user.UserDB;
+import org.burrow_studios.obelisk.server.db.entity.BoardDB;
+import org.burrow_studios.obelisk.server.its.IssueState;
+import org.burrow_studios.obelisk.server.db.entity.ProjectDB;
+import org.burrow_studios.obelisk.server.moderation.ProjectState;
+import org.burrow_studios.obelisk.server.db.entity.TicketDB;
+import org.burrow_studios.obelisk.server.moderation.TicketState;
+import org.burrow_studios.obelisk.server.db.entity.GroupDB;
+import org.burrow_studios.obelisk.server.db.entity.UserDB;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
