+++ obelisk-server/src/main/java/org/burrow_studios/obelisk/server/db/entity/BoardDB.java
@@ -1,7 +1,8 @@
-package org.burrow_studios.obelisk.server.its.db;
+package org.burrow_studios.obelisk.server.db.entity;
 
 import com.google.gson.JsonObject;
 import org.burrow_studios.obelisk.server.db.DatabaseException;
+import org.burrow_studios.obelisk.server.its.IssueState;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Set;
