+++ obelisk-server/src/main/java/org/burrow_studios/obelisk/server/db/entity/GroupDB.java
@@ -1,4 +1,4 @@
-package org.burrow_studios.obelisk.server.users.db.group;
+package org.burrow_studios.obelisk.server.db.entity;
 
 import com.google.gson.JsonObject;
 import org.burrow_studios.obelisk.server.db.DatabaseException;
