+++ obelisk-server/src/main/java/org/burrow_studios/obelisk/server/its/IssueState.java
@@ -1,4 +1,4 @@
-package org.burrow_studios.obelisk.server.its.db;
+package org.burrow_studios.obelisk.server.its;
 
 public enum IssueState {
     OPEN,
